silence warnings
The main cause of the warnings was that we were not checking explicitly
for impossible cases when doing pattern matching. For example,
a grammar rule may specify that only three quantifier operators exist
(*, +, ?) but the compiler doesn't know this and complains because
we don't have a case for any string other than "*", "+", and "?".  We added
an explicit case for handling these (impossible) situations that throws an exception,
which should never happen.